Added LaserDectFunct.py
A function which takes in an image and returns an (x,y) pair of the
average of all the points in the image which have been confirmed to be
red ([255, 0, 0]).

Includes basic code to run with webcam. I have no laser pointer, so I
can only attest that it works well with an image I created to test its
ability to discern different colors